requirements.txt
pykka
pySqsListener